requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
jinja2>=3.1.0
python-dotenv>=1.0.0
praw>=7.7.0
anthropic>=0.39.0
posthog>=3.5.0
pydantic>=2.6.0
pytest>=8.0.0
httpx>=0.27.0
